Modified   Engine/Binaries/Frontend/Frontend.exe Modified   Engine/Frontend/Frontend/System/Build/EngineBuilder.cs Modified   Engine/Source/Developer/ResourceSystem/ResourceAssetParser/TTFAssetParser/TTFAssetParser.ModuleRules.cs Modified   Engine/Source/Programs/ReflectionTool/Main.cpp Modified   Engine/Source/Programs/ReflectionTool/ReflectionToolAllocators.cpp Modified   Engine/Source/Runtime/Containers/Buffer.h Modified   Engine/Source/Runtime/MemoryManagement/Allocators/AllocatorBase.cpp Modified   Engine/Source/Runtime/MemoryManagement/Allocators/AllocatorBase.h Modified   Engine/Source/Runtime/MemoryManagement/Allocators/Allocators.ModuleRules.cs Modified   Engine/Source/Runtime/MemoryManagement/Allocators/CustomAllocator.cpp Modified   Engine/Source/Runtime/MemoryManagement/Allocators/CustomAllocator.h Modified   Engine/Source/Runtime/MemoryManagement/Allocators/DefaultAllocator.cpp Modified   Engine/Source/Runtime/MemoryManagement/Allocators/DefaultAllocator.h Modified   Engine/Source/Runtime/MemoryManagement/Allocators/DynamicSizeAllocator.cpp Modified   Engine/Source/Runtime/MemoryManagement/Allocators/DynamicSizeAllocator.h Modified   Engine/Source/Runtime/MemoryManagement/Allocators/FixedSizeAllocator.cpp Modified   Engine/Source/Runtime/MemoryManagement/Allocators/FixedSizeAllocator.h Modified   Engine/Source/Runtime/MemoryManagement/Allocators/FrameAllocator.cpp Modified   Engine/Source/Runtime/MemoryManagement/Allocators/FrameAllocator.h Missing    Engine/Source/Runtime/MemoryManagement/Allocators/Initializer.cpp Missing    Engine/Source/Runtime/MemoryManagement/Allocators/Initializer.h Missing    Engine/Source/Runtime/MemoryManagement/Allocators/MemoryHeader.h Modified   Engine/Source/Runtime/MemoryManagement/Allocators/RootAllocator.cpp Modified   Engine/Source/Runtime/Platform/PlatformMemory.cpp Modified   Engine/Source/Runtime/Platform/PlatformMemory.h Modified   Engine/Source/Runtime/RenderSystem/RenderDevice/DirectX12RenderDevice/Private/DirectX12Wrapper.h Modified   Engine/Source/Runtime/ResourceSystem/ResourceAssetParser/FontAssetParser/FontAssetParser.ModuleRules.cs Modified   Engine/Source/Runtime/ResourceSystem/ResourceAssetParser/FontAssetParser/FontParser.cpp Modified   Engine/Source/Runtime/ResourceSystem/ResourceAssetParser/TextureAssetParser/TextureParser.cpp Modified   Engine/Source/Runtime/ThirdParty/GLSLANG/GLSLANG.ModuleRules.cs Modified   Engine/Source/Test/TestDockable/TestDockable.cpp Modified   Engine/Source/Test/TestLauncher/TestLauncher.cpp Modified   Engine/Source/Test/TestParallelizing/TestParallelizing.cpp Added      Engine/Source/Runtime/MemoryManagement/Allocators/Private/MemoryHeader.h

+++ Engine/Source/Runtime/RenderSystem/RenderDevice/DirectX12RenderDevice/Private/DirectX12Wrapper.h
@@ -149,7 +149,7 @@ namespace Engine
 						FrameAllocator allocator("Debug Message Allocator", RenderSystemAllocators::RenderSystemAllocator, MegaByte);
 						for (uint64 i = 0; i < count; ++i)
 						{
-							uint64 size = 0;
+							SIZE_T size = 0;
 							InfoQueue->GetMessage(i, nullptr, &size);
 
 							D3D12_MESSAGE* message = ReinterpretCast(D3D12_MESSAGE*, AllocateMemory(&allocator, size));
